requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain
langchain-community
pypdf
sentence-transformers
